Remove Info and Activity tabs when viewing single ERC1155 token instance #3833
