+++ AlphaWallet/Tokens/ViewModels/TokenInstanceViewModel.swift
@@ -1,8 +1,23 @@
-// Copyright © 2018 Stormbird PTE. LTD.
+//
+//  TokenInstanceViewModel2.swift
+//  AlphaWallet
+//
+//  Created by Vladyslav Shepitko on 07.09.2021.
+//
 
-import Foundation
-import BigInt
 import UIKit
+import BigInt
+
+enum TokenInstanceViewConfiguration {
+    case header(viewModel: TokenInfoHeaderViewModel)
+    case field(viewModel: TokenInstanceAttributeViewModel)
+    case attributeCollection(viewModel: OpenSeaOpenSeaAttributeCollectionViewModel)
+}
+
+enum TokenInstanceViewMode {
+    case preview
+    case interactive
+}
 
 struct TokenInstanceViewModel {
     let tokenId: TokenId
@@ -11,6 +26,7 @@ struct TokenInstanceViewModel {
     let assetDefinitionStore: AssetDefinitionStore
     private let displayHelper: OpenSeaNonFungibleTokenDisplayHelper
     var backgroundColor: UIColor = Colors.appBackground
+    private let tokenHolderHelper: TokenInstanceViewConfigurationHelper
 
     init(tokenId: TokenId, token: TokenObject, tokenHolder: TokenHolder, assetDefinitionStore: AssetDefinitionStore) {
         self.tokenId = tokenId
@@ -18,6 +34,7 @@ struct TokenInstanceViewModel {
         self.tokenHolder = tokenHolder
         self.assetDefinitionStore = assetDefinitionStore
         self.displayHelper = OpenSeaNonFungibleTokenDisplayHelper(contract: tokenHolder.contractAddress)
+        self.tokenHolderHelper = TokenInstanceViewConfigurationHelper(tokenId: tokenId, tokenHolder: tokenHolder)
     }
 
     var actions: [TokenInstanceAction] {
@@ -46,126 +63,16 @@ struct TokenInstanceViewModel {
         }
     }
 
-    var supplyModelViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.supplyModel.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeSupplyType(), attributedValue: $0)
-        }
-    }
-
-    var transferableViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.transferable.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeTransferable(), attributedValue: $0)
-        }
-    }
-
-    var meltValueViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.meltStringValue.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeMelt(), attributedValue: $0)
-        }
-    }
-
-    var meltFeeRatioViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.meltFeeRatio.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString(String($0))
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeMeltFeeRatio(), attributedValue: $0)
-        }
-    }
-
-    var meltFeeMaxRatioViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.meltFeeMaxRatio.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString(String($0))
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeMeltFeeMaxRatio(), attributedValue: $0)
-        }
-    }
-
-    var totalSupplyViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.totalSupplyStringValue.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeTotalSupply(), attributedValue: $0)
-        }
-    }
-
-    var circulatingSupplyViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.circulatingSupplyStringValue.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeCirculatingSupply(), attributedValue: $0)
-        }
-    }
-
-    var reserveViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.reserve.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeReserve(), attributedValue: $0)
-        }
-    }
-
-    var nonFungibleViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.nonFungible.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString(String($0))
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeNonFungible(), attributedValue: $0)
-        }
-    }
-
-    var availableToMintViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.mintableSupply.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString(String($0))
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeAvailableToMint(), attributedValue: $0)
-        }
-    }
-
-    var issuerViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.issuer.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString(String($0))
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeIssuer(), attributedValue: $0)
-        }
-    }
-
-    var transferFeeViewModel: TokenInstanceAttributeViewModel? {
-        return tokenHolder.values.transferFee.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString(String($0))
-        }.flatMap {
-            .init(title: R.string.localizable.semifungiblesAttributeTransferFee(), attributedValue: $0)
-        }
-    }
-
-    var createdDateViewModel: TokenInstanceAttributeViewModel {
-        let string: String? = tokenHolder.values.collectionCreatedDateGeneralisedTimeValue?.formatAsShortDateString()
-        let attributedString: NSAttributedString? = string.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0)
-        }
-        return .init(title: R.string.localizable.semifungiblesCreatedDate(), attributedValue: attributedString)
-    }
-
-    var descriptionViewModel: TokenInstanceAttributeViewModel {
-        let string = tokenHolder.values.collectionDescriptionStringValue.flatMap {
-            TokenInstanceAttributeViewModel.defaultValueAttributedString($0, alignment: .left)
-        }
-        return .init(title: nil, attributedValue: string, isSeparatorHidden: true)
-    }
-
     var tokenIdViewModel: TokenInstanceAttributeViewModel? {
-        guard let values = tokenHolder.values(tokenId: tokenId), !values.isEmpty else { return nil }
-        return values.tokenIdStringValue.flatMap { tokenId in
-            .init(title: R.string.localizable.semifungiblesTokenId(), attributedValue: TokenInstanceAttributeViewModel.defaultValueAttributedString(tokenId))
-        }
+        tokenHolderHelper.tokenIdViewModel
+    }
+
+    var tokenImagePlaceholder: UIImage? {
+        return R.image.tokenPlaceholderLarge()
     }
 
     var configurations: [TokenInstanceViewConfiguration] {
-        guard let values = tokenHolder.values(tokenId: tokenId), !values.isEmpty else { return [] }
+        guard let values = tokenHolderHelper.values else { return [] }
 
         var previewViewModels: [TokenInstanceViewConfiguration] = []
         if let viewModel = tokenIdViewModel {
@@ -175,30 +82,54 @@ struct TokenInstanceViewModel {
         }
 
         previewViewModels += [
-            issuerViewModel,
-            transferFeeViewModel,
-            createdDateViewModel,
-            meltValueViewModel,
-            meltFeeRatioViewModel,
-            meltFeeMaxRatioViewModel,
-            totalSupplyViewModel,
-            circulatingSupplyViewModel,
-            reserveViewModel,
-            nonFungibleViewModel,
-            availableToMintViewModel,
-            transferableViewModel,
+            tokenHolderHelper.issuerViewModel,
+            tokenHolderHelper.transferFeeViewModel,
+            tokenHolderHelper.createdDateViewModel,
+            tokenHolderHelper.meltValueViewModel,
+            tokenHolderHelper.meltFeeRatioViewModel,
+            tokenHolderHelper.meltFeeMaxRatioViewModel,
+            tokenHolderHelper.totalSupplyViewModel,
+            tokenHolderHelper.circulatingSupplyViewModel,
+            tokenHolderHelper.reserveViewModel,
+            tokenHolderHelper.nonFungibleViewModel,
+            tokenHolderHelper.availableToMintViewModel,
+            tokenHolderHelper.transferableViewModel,
         ].compactMap { each -> TokenInstanceViewConfiguration? in
             return each.flatMap { TokenInstanceViewConfiguration.field(viewModel: $0) }
-        }
+        } 
 
         let value: BigInt = values.valueIntValue ?? 0
+        let attributedValue = TokenInstanceAttributeViewModel.defaultValueAttributedString(String(value))
         previewViewModels += [
-            .field(viewModel: .init(title: R.string.localizable.semifungiblesValue(), attributedValue: TokenInstanceAttributeViewModel.defaultValueAttributedString(String(value))))
+            .field(viewModel: .init(title: R.string.localizable.semifungiblesValue(), attributedValue: attributedValue))
         ]
+
         if let description = values.descriptionAssetInternalValue?.resolvedValue?.stringValue.nilIfEmpty {
+            let attributedValue = TokenInstanceAttributeViewModel.defaultValueAttributedString(description, alignment: .left)
             previewViewModels += [
                 .header(viewModel: .init(title: R.string.localizable.semifungiblesDescription())),
-                .field(viewModel: .init(title: nil, attributedValue: TokenInstanceAttributeViewModel.defaultValueAttributedString(description, alignment: .left), isSeparatorHidden: true))
+                .field(viewModel: .init(title: nil, attributedValue: attributedValue, isSeparatorHidden: true))
+            ]
+        }
+
+        if !tokenHolderHelper.attributes.isEmpty {
+            previewViewModels += [
+                .header(viewModel: .init(title: R.string.localizable.semifungiblesAttributes())),
+                .attributeCollection(viewModel: .init(attributes: tokenHolderHelper.attributes))
+            ]
+        }
+
+        if !tokenHolderHelper.stats.isEmpty {
+            previewViewModels += [
+                .header(viewModel: .init(title: R.string.localizable.semifungiblesStats())),
+                .attributeCollection(viewModel: .init(attributes: tokenHolderHelper.stats))
+            ]
+        }
+
+        if !tokenHolderHelper.rankings.isEmpty {
+            previewViewModels += [
+                .header(viewModel: .init(title: R.string.localizable.semifungiblesRankings())),
+                .attributeCollection(viewModel: .init(attributes: tokenHolderHelper.rankings))
             ]
         }
